Update ktfmt (#156)
Per [b/343255536](https://b.corp.google.com/issues/343255536),

This update ktfmt to the latest version. This is needed to support a kmp
migration, and the formatting changes from `0.16.0` to `0.18.0` are
enough to pollute the kmp pr. So this change is being done separately to
help reduce the diff in kmp related changes.